backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
alembic==1.15.2
psycopg[binary]==3.2.7
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
pydantic-settings==2.9.1
email-validator==2.2.0

